+++ backend/src/index.ts
@@ -7,7 +7,6 @@ import config from "./configs/config";
 // Route
 import AuthRoute from "./routes/AuthRoute";
 import ReportRoute from "./routes/ReportRoute";
-import { LikelihoodRoute } from "./routes/LikelihoodRoute";
 import MedErrorRoute from "./routes/MedErrorRoute";
 // Plugin
 import { DBMain, DBSec } from './plugins/db'
@@ -28,7 +27,6 @@ const app = new Elysia()
   .use(AuthRoute)
   .use(MedErrorRoute)
   .use(ReportRoute)
-  .use(LikelihoodRoute)
   .get('/alive', () => ({
     statusCode: StatusCodes.OK,
     status: 'alive',